add notes, fix timezone
Mapped the Nike+ note to Runkeeper's note.

Nike+ activities are in UTC, whereas Runkeeper wants them in local
time; this was resulting in Runkeeper run start times being off by
several hours (since I'm at -5:00 UTC).  Hack around this by
subtracting the local gmt offset.  It would be better to calculate the
gmt offset of the gps location, but this is good enough for me.